feat: Add character name selection and display in shop scene
This commit adds the functionality to select a character name randomly from a list of choices (Gary, Adam, Eric, James) and display it in the shop scene. The player's character name is set in the `set_player_name()` function called during `_ready()`. The shop scene's RichTextLabel now displays a message that includes the selected character name.